server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
pdfminer.six==20231228
reportlab==4.2.2
pillow==10.4.0
openai==1.58.1
python-dotenv==1.2.0
docling==2.58.0
